Clean up usage of data attributes in list_session_groups. (#6503)
This is a cleanup PR to list_session_groups.py, reducing its reliance on
data attributes.

We now keep the set of data attributes small - to correspond exactly to
the arguments of the `__init__` method (`self._request_context`,
`self._backend_context`, `self._experiment_id`, and `self._request`).

Two motivating factors:
* Recently we have split list_session_groups.py into two major branches
of logic: The branch that generates "classic" hparams from tensor
metadata; and the branch that generates hparams from the data provider.
Many of the existing data attributes applied only to the first branch
and, in some cases, were not strictly guaranteed to be set since they
are no longer set in `__init__`.
* We can now move `self._extractors` and `self._filters` out of
the`__init__` method now that we know they won't apply to the data
provider-based hparams. It means no calculation for the "classic"
hparams logic is necessary in the `__init__`.