Change to facilitate Angular Material theming API
We're in the process of changing the Angular Material components to use CSS variables for theming. This involves moving some declarations from the theme to the component base styles which changes its specificity. One such change that breaks the Tensorboard screenshot tests is https://github.com/angular/components/pull/27479.

The problem boils down to the fact that because the styles that override the form field color are in the base styles now _and_ the Tensorboard dark theme targets `body.dark-mode`, only in dark mode some of the select declarations aren't being overwritten anymore. This won't be a problem once Tensorboard moves away from the legacy Angular Material components, but for now this patch is necessary to ensure that it doesn't blend into the toolbar.

**Note:** these changes will require some minor updates to the existing internal screenshots. See cl/549228851 for more info.