Bug Fix: Prevent custom modal from being closed until after it is opened (#6523)
## Motivation for features / changes
In #6504 I introduced a bug to the column selector modal when it was
opened via the + button. The effect was that the modal was being closed
in the same event that opened it. Rather than changing the way that
specific event was being dispatched I instead am preventing the modal
from being closed until at least a single frame has been rendered.

## Screenshots of UI changes (or N/A)
It didn't work before 

![3f02da9d-148b-44d8-bc1b-7951bb789808](https://github.com/tensorflow/tensorboard/assets/78179109/77ee2fc6-8a1b-4280-a224-20055ce6d488)

It works now

![a17b2e80-40c6-4bd4-accf-40ee5cdd4b47](https://github.com/tensorflow/tensorboard/assets/78179109/31aab2fd-ad1b-4d04-ac82-731c737a0627)