Chore: Remove header attribute of sortingInfo interface (#6524)
## Motivation for features / changes
This is the final change in a long line of changes to move to using
name. The original and functional change is #6362. This had to add name
as an optional parameter only in order to avoid internal breakages. Now
that we had name we could add the use of name internally(Googlers see
cl/534191646) . Then we could switch to make the header optional and the
name required which was done here: #6400. After header was optional we
removed the use of headers internally(googlers see cl/552643471).
Finally in this PR we removed the header attribute.