Throttle tags request (#6527)
## Motivation for features / changes
For years we have been making two requests to fetch tags then canceling
the first one. This occurs because the request is dispatched when either
the dashboard is shown without data or a new plugin is loaded. When the
dashboard is initially loaded both of these conditions are true.

## Screenshots of UI changes (or N/A)
Before:

![image](https://github.com/tensorflow/tensorboard/assets/78179109/cbaffa77-4388-4222-bf84-fd93494f00b6)

After:

![image](https://github.com/tensorflow/tensorboard/assets/78179109/78f13434-c0ee-402f-bf7c-5dc5ea05d0d1)