Updates import and BUILD dependency in test from example (#6531)
## Motivation for features / changes

I expect the behavior for relative imports changed after the python rule
changed in #6495, and made this test fail, so it now needs to be an
absolute import. (Unsure of why this was not making CI fail... maybe
this particular target is not run as part of CI, since it's just an
example.)

## Technical description of changes

Adds a BUILD target for the submodule that was previously imported with
a relative import. With the BUILD target and the dependency, the module
can now be referenced as an absolute import.

## Screenshots of UI changes (or N/A)

N/A

## Detailed steps to verify changes work correctly (as executed by you)

Tests that were previously failing now pass.

## Alternate designs / implementations considered (or N/A)

N/A