Change return type of list_hyperparameters() operation. (#6537)
Change the return type of DataProvider.list_hyperparameters() to be a
complex "Result" type that can contain multiple pieces of information.

Where previously it was a Collection[Hyperparameter], it is now a
ListHyperparametersResult. It contains a "hyperparameters" field that
contains the same data as the previous return type. We also add a
"session_groups" field for future usage.

We change the hparams plugin to be able to handle both the old and the
new return types.