Hparams: Fix metric info generation for sessions without run names. (#6541)
Fix a bug where we couldn't generate metric infos for
HyperparameterSessionRun without run names. We were failing to match the
sessions with the runs returned by scalars_metadata().

As an example:

When a HyperparameterSessionRun does not contain `run` field, we were
generating session_names of the form 'exp1/', 'exp2/', etc.. with a
trailing '/'. Meanwhile, runs would be just of the form 'exp1/run_name'.
The logic to match paths in _find_longest_parent_path() would first try
to find 'exp1/run_name' in session_names and then 'exp1' in
session_names. The second try would fail because of the trailing slashes
in session_names.

So, instead, when there is no `run` field in a HyperparameterSessionRun,
we drop the final '/' and just generate names like 'exp1', 'exp2',
etc... and the algorithm in _find_longest_parent_path() succeeds.