HParams: Create filter dialog component (#6493)
## Motivation for features / changes
We want to be able to filter the runs table by hparam. The existing
logic for this is handled by a bunch of inner components found in the
runs_table_component which makes it hard to share. In this PR I am
replicating that UI with some enhancements in a standalone component
which will later be used by the data_table_component.

See #6488 for screenshots and a more complete description of what is
being added

## Alternate designs / implementations considered (or N/A)
I'm not sure that this component should be part of the data table but
that is simplest place to put it for now.