Hparams: Allow sessions without name to pull all run,tag combinations as metrics. (#6546)
Motivation: Sometimes a user is viewing a single experiment and the
DataProvider determines that experiment contains exactly one session. In
this case the DataProvider will return a session with both empty
experiment_id and empty run.

We want this "empty-name session" to represent all runs in the
experiment. But we have a bug where the /experiment and /session_group
responses do not contain any metrics for the session. We want to instead
designate all (run,tag) combinations as separate metrics to include in
the /experiment and /session_groups responses.

The first change to get this to work:

* When generating session names, a session with experiment_id="" and
run="" should generate an empty session_name "" (instead of using the
somewhat-meaningless input experiment id as the session_name).
 
The side-effect of the first change is that calls to
_find_longest_parent_path() will return "" for all runs passed into it -
effectively saying that all runs belong to the "empty-name session". So
the second change to get this to work:

* When _find_longest_parent_path() returns "" instead of None, treat
this as a match with "empty-name session" instead of ignoring it.