Bug Fix: Start using expect_angular_legacy_material_checkbox (#6547)
## Motivation for features / changes
#6493 Broke our internal sync Googlers see
[cl/557499468](http://cl/557499468)

The issue is related to our internal Angular component migration. New
components are not allowed to rely on the legacy components. Due to this
the new component added in #6493 uses the non legacy checkbox. However,
our dependency remapping always remaps dependencies on material_checkbox
to legacy_material_checkbox.

To fix this I am changing the name of the remapping rule in
[cl/557515881](http://cl/557515881) and adjusting all the OSS usage
EXCEPT for the usage in filter_dialog.

[cl/557590954](http://cl/557590954) is also required to map the testing
dependency needed to use `MatCheckboxHarness`