Hparam: Add selectors to retrieve data written by the new hparams effects (#6544)
## Motivation for features / changes
This is a continuation of #6540
Now that data is being retrieved from the hparams plugin and written to
the store I am adding a pair of selectors to retrieve it and inject it
into the the data retrieved from `getRuns` and `getRunsForExperimentId`.

The logic used to translate `SessionGroups` to `Record<RunId/Group,
Value>` is very similar to that being used by the old code in
`runs_data_source`. I could imagine this being refactored to retrieve
the run/group relationships from another part of the store in the future
in order to better integrate with our plans to stop relying on the
metrics portion of the hparams payload.

## Note For Reviewer(s)
A large majority of the lines changed in this PR are indent changes
resulting from changing the method used to mock the state during tests.

I have opted not to add additional test to the runs or common selectors
(though I did have to fix 16) as the intention is for this to be a drop
in replacement with no new features added. The number of test failures
which I had to fix left me feeling confident of the existing test
coverage.