HParams: Add new state property to store hparam and metric filters (#6553)
## Motivation for features / changes
Now that the filter dialog is merged #6493 we need to add support for
filtering by hparams. However, #6544 changed the way hparam values are
read. This change adds a new property to state in which to store
dashboard related hparam and metric filters.

See #6488 for the WIP integration with the runs_table_container +
tb_data_table -> common_selectors + hparam_selectors.