Bug Fix: Update the port scanning logic to properly skip ports in use (#6560)
## Motivation for features / changes
#6552

## Technical description of changes
The port scanning logic originally added in #1851 does not seem to work
anymore. WerkzeugServer no longer raises an Error when attempting to
bind to a port which is in used and instead handles the issue silently.

The current Werkzeug version is 2.3.7 but I had 2.3.4 installed when I
first verified this. I tried installing 2.2.0 and the issues was not
present so I assume it was introduced in 2.3.0.
https://pypi.org/project/Werkzeug

## Detailed steps to verify changes work correctly (as executed by you)
1) Start TensorBoard, note the port it is bound to (it should be 6006)
2) In a separate terminal start TensorBoard again.
3) TensorBoard should successfully bind to a different port (this time
6007)