Suggested Cards: Create data source for persisting card interactions to localStorage (#6439)
## Motivation for features / changes
We are working on a feature to suggest cards to users at the top of the
timeseries dashboard. Suggestions are generated based on previous
interactions with TensorBoard. In order to determine what those previous
interactions were we are storing them to localStorage.
See #6437 for more information.
Googlers see [go/tb-suggested-cards](go/tb-suggested-cards)