HParams: Add effect to remove hparam filters when the related column is removed (#6572)
## Motivation for features / changes
We found it confusing that a user could remove an hparam column but
still have related filters in place. This also lead to awkward UX where
they would need to re-add the column in order to remove the filter.

## Technical description of changes
Since the action is a runs action I added an effect to runs effects
which dispatches a newly created action to clear out related hparam
filters.