Bug Fix: update the range input widget to only show the slider when min and max are different (#6591)
## Motivation for features / changes
When teamfooding the hparams in timeseries feature we discovered that
the interval filter widget doesn't look good when there is only a single
value. This is actually due to the `tb-range-input` widget looking bad
when the min and the max are equal. I'm opting to simply not show the
slider when `min === max`

For Googlers:
https://screenshot.googleplex.com/5rZbjsd6ijp8dkY