HParams: Make icons buttons in data table header (#6594)
## Motivation for features / changes
To allow for tabbing to buttons and for general best practice clickable
elements should be buttons. So the 3 dot menu button was changed into a
button instead of a div. For sorting the entire header is clickable. We
did not want to make the entire header a button so I turned the sorting
icon into a button.

I also made some styling changes to spread out the icons and make them
easier to click.

## Technical description of changes
When I made the sorting icon a button I had to add the click event to
make it sort. I also had to stop propagation so the header's click event
was not fired.

## Screenshots of UI changes (or N/A)
![2023-09-22_11-05-35
(1)](https://github.com/tensorflow/tensorboard/assets/8672809/29b33b4c-6122-49fd-b299-b55a12d69011)