HParams: Do not scroll regex filter in the horizontal direction (#6596)
## Motivation for features / changes
I originally started to fix a bug where the bottom border in the regex
did not stretch to the end of the table once scrolled. However, I
decided there is no reason to scroll the regex at all and it can
actually be kind of annoying. So I change the x scroll to just be the
table

## Screenshots of UI changes 
![2023-09-22_16-57-36
(1)](https://github.com/tensorflow/tensorboard/assets/8672809/049cc341-6787-48d8-bb21-1e894434cf47)
(or N/A)