Hparams: Generate map of runs to hparams by matching with session names. (#6600)
The way we generate the mapping of runs to hparams for the dashboard
table must be changed.

The primary reason for this is that the current algorithm relies on
metric names from the session_group call to match runs with sessions
before mapping the session's hparams to the run. But this is not always
an accurate way of identifying the complete set of runs that belong to a
session and, besides, we will be turning off metric retrieval for
dashboard's session_group call in a change later this week.

We instead rely on the property of sessions that their name is the
prefix for all runs that belong to the session.

Note: The new algorithm requires getting all run ids from the runs state
so we move the algorithm from hparams_selectors to runs_selectors in
order to avoid a circular BUILD dependency.

Note: Some of the comments in this change claim that the hparams data
source does not retrieve metrics data. This is not true, yet, but will
be by the end of the week.