Give fixed width to runs table's checkbox and color columns. (#6637)
## Motivation for features / changes

The checkbox (aka 'selected') and color icon columns in the new runs
data table would have variable width in the table's flex layout. This
led to some awkward looking spacing for these columns.

We want to fix the width of these columns.

## Technical description of changes

Similar to strategies for other tables (mat-table and the old runs
table, for example), add a column-specific class to each cell and
customize the width of the 'selected' and 'color' columns using these
classes.