Adds hparam filterbar to runs table (#6692)
## Motivation for features / changes
To view applied hparam filters in the runs table.

## Technical description of changes
If a user filters the runs table based on a hparam column, the filter
will be shown as a chip above the table. Clicking the chip will allow
the user to modify the filter via a dialog. Clicking the close button on
the chip will remove the filter.

## Screenshots of UI changes (or N/A)
<img width="871" alt="Screenshot 2023-12-04 at 18 30 27"
src="https://github.com/tensorflow/tensorboard/assets/736199/d3acc72b-4526-4590-b390-3442fe435b7d">