tbdev turndown: add export converter script (#6698)
This adds a standalone script `convert_tbdev_export.py` that can convert
the JSON-based export format for TensorBoard.dev (produced by
`tensorboard dev export`) into tfevents files that can be used to
display the exported data in open source TensorBoard. See documentation
in the script itself for more details about how it's intended to be
used.

Note that I've structured this as a standalone script in the style of
diagnose_tensorboard.py (i.e. no BUILD target, no dependencies, no unit
tests). Arguably, had we done this earlier, it would be better to have
integrated this into the `tensorboard dev export` command itself as an
output format option. But we're planning to remove the `tensorboard dev`
command entirely once the turndown happens at EOY, so keeping it
separate ensures that we can still link to this script for a while as a
resource for anyone who has already exported their experiments to do
conversions post-turndown.

Tested: did manual testing with a range of my own experiments, including
some with multiple runs, different data types (scalars, tensors, blobs),
etc. Confirmed that the conversion completed without errors and that it
looked as expected in the TensorBoard UI. I've omitted unit tests
because the manual end-to-end tests are what we ultimately care more
about, and I don't expect any further significant development to happen
for this script.