Chore: Fix an issue with property renaming in polymer (#6401)
## Motivation for features / changes
We keep the property renaming compiler flag turned off because of the
many issues it causes with polymer. This should bring us one step closer
towards enabling it.
b/281726974

Googlers, see that this imports correctly cl/585788456

## Technical description of changes
The hack of creating custom elements and then referencing properties on
them doesn't really work with property renaming turned on because the
portions of the code which reference the properties are often not being
changed. To get around this while still maintaining the binary
seperation I am updating the code to write properties to `window`

## Screenshots of UI changes (or N/A)
Properties are not being renamed

![image](https://github.com/tensorflow/tensorboard/assets/78179109/863f5cd7-4805-4e8f-bf82-309d3dd51716)