Bump webtesting rules version.
We haven't updated this dependency on the bazel webtesting rules for a
while, and it looks like we need to do it to be able to also update our
reference to the rules_closure target, to unblock #6721.

To be clear, this PR is a prerequisite to pin the rules_closure version
in a separate PR, which in turn will unblock #6721.

An attempt to use the versioned browsers provided by the
rules_webtesting dependency was created in #6772, but some tests were
failing. It seems that some components or tests were written in a way
that depends on the specific browser (?), so to be able to update this
dependency, this continues to use the same browser version, only
updating the rules we use to define our testing browser, to be
compatible with a newer version of rules_webtesting.