Hparams: Workaround problem with '/.' in run names. (#6791)
Some responses from data_provider.read_last_scalars() may contain run
names of the following form: 'SOME_EXPERIMENT_OR_SOMETHING/.', with a
trailing '/.' at the end.

'.' always behaves sort of weird with Hparams and this is no exception.
The problem in this case is that sessions of this sort could not be
matched with metric values. No metric values would appear for them in
the hparams dashboard.

This is because the logic to generate metric names drops the '/.' and so
we could not match. That code relies on some os-level path operations
and we know that '.' has special meaning for filesystems:

https://github.com/tensorflow/tensorboard/blob/23073c55c03f2f1a9da8c7e0bb9db3349dc15c90/tensorboard/plugins/hparams/metrics.py#L39

We fix the problem by transforming the result of
data_provider.read_last_scalars() to apply the same os-level path
operations to the run names (ie the keys of the result).