Hparams plugin: Retrieve and render only first 1000 hparams. (#6807)
We limit the number of hparams retrieved and rendered in the Hparams
plugin. Otherwise the plugin can become very slow and even crash.

Unlike #6780, we do not provide an option for loading more in the
Hparams plugin. We may provide this option later if there is demand. But
generally we think 1000 should be enough, especially if the underlying
hparams provider is sorting the results with "differs" fields first.


![image](https://github.com/tensorflow/tensorboard/assets/17152369/da1f2c26-3804-4faf-be4a-03bb8aaa178d)