Makes data table add column sticky (#6839)
## Motivation for features / changes
The add column button in the runs table will be hidden when the table is
wide, which will often be the case. This makes the "add" feature very
difficult to discover.

context: b/332788091

## Technical description of changes
- Moves the add button column outside of and adjacent to the data table
(previously it was a column of the table) and makes it sticky
- Adds additional configuration options to data table to allow runs and
scalar tables to have slightly different designs

## Screenshots of UI changes (or N/A)
Notice the add column is fixed to the right:

![sticky](https://github.com/tensorflow/tensorboard/assets/736199/0f7aca5c-0903-438a-bafd-ddc033e262f7)

Scalar table add column is slightly narrower:

![image](https://github.com/tensorflow/tensorboard/assets/736199/e8c37505-4c3f-41bd-98f3-3b31f99e206e)

## Alternate designs / implementations considered (or N/A)
- Tried adding the add columns to the consumer components of the data
table (runs data table and scalar card data table) instead. But this led
to too much code duplication in the components and templates.
- Tried making the add column a regular table column, instead of adding
it outside in a separate div. But making the column sticky requires
styling the table rows, which data table doesn't provide access to (rows
are configured solely in the consumer components). Also, the chosen
implementation is arguably more consistent with the current design as
the add "column" doesn't actually function as a table column at all