Fix the issue with coloring when trying to add more experiments.
There was a problem when user tried to 1) enable coloring by experiment
name, 2) trying to add more experiments. It stemmed from not
implementing the serialization and deserialization of REGEX_BY_EXP in
the query pararms of the route. To followup the #6846 and #6847 I have
added the 'regex_exp:' as a query parameter for coloring by experiment
name.