Fix the issue with coloring when trying to add more experiments. (#6861)
There was a problem when user tried to 1) enable coloring by experiment
name, 2) trying to add more experiments. It stemmed from not
implementing the serialization and deserialization of REGEX_BY_EXP in
the query pararms of the route. To followup the #6846 and #6847 I have
added the 'regex_exp:' as a query parameter for coloring by experiment
name.

## Motivation for features / changes

To fix the issue with the new feature. 

## Technical description of changes

Added `regex_exp:` query param for serialization and deserialization of
`REGEX_BY_EXP` GroupBy. It ensures that whenever user share the URL the
filter will remain in place.

## Screenshots of UI changes (or N/A)

## Detailed steps to verify changes work correctly (as executed by you)

1) Enable the coloring by experiment name.
2) Try to add more experiments via pressing `Add more experiments` in
the top panel.

## Alternate designs / implementations considered (or N/A)