Explicitly set `standalone` for all Angular components (#6912)
Angular v19 is going to change the default value for `standalone` from
`false` to `true`. Even though tensorboard is on an older version of
Angular on GitHub, the version inside Google runs at HEAD. We're in the
process of changing existing code in google to explicitly set
`standalone: false` for existing code, so I'm sending this change to the
source of truth here.