Removes duplicate style specs. (#6984)
## Motivation for features / changes
Angular v15, which is the one currently used in the OSS repo, uses
styles with format `--mdc-` to style material components.

Later versions of angular migrated to a `--mat-` prefix, and Angular v20
will update some of the names again.

Our internal codebase uses the styles defined internally, so we need to
use the new format.

We recently set up translations internally from these "old" `--mdc-`
stylings to the newer `--mat-` names, so we don't need them duplicate
here, which could potentially be a source for confusion.

Googlers, see cl/744892387 and cl/744895577 for some details.

## Technical description of changes
Remove some scss lines that are unused in the public repo.