Remove deprecated call to flatDoc.normalize (#7011)
## Motivation for features / changes
This was a deprecated call to a no-op function. The function has now
been removed in a newer version of the library. We need to support this
newer library version to unblock an internal lsc Googlers see
cl/771125806

I also needed to add a new source for the jsoup dependency. Previously
this was being provided by the operating system (CI currently uses
ubuntu 22.04 lts), but that version is now 3 years out of date.