Updates (vendored) bleach to 6.2.0 and removes dependencies on `six` and `html5lib`
Addresses #7010.

The old version of bleach that we were using, needed dependencies `six`
and `html5lib` to be installed, and the old version of "six" that we
were using for CI via bazel is not supported in python >= 3.12, which
causes some issues at least for local builds and CI when using newer
versions of python.